Clarify max CPUs (#832)
* Clarify max CPUs

Historically Cirrus only supported up to 8 CPUs instances on community clusters. When [CPU-based limits](https://medium.com/cirruslabs/new-cpu-based-limits-for-cirrus-ci-b0dcd82f4615) were introduced it appeared that it became possible to request 16.0 cores. But it wasn't desired since all the VMs in the cluster are tuned for running containers with 8.0 cores or less. Recently there was uptick of people using only 16.0 cores containers, there were so many of them (almost 100 repositories committing pretty frequently) that it started affecting scheduling times, auto-scaling and utilization all together in not a good way.

This PR clarifies the limit which will now be enforced in order to guarantee speedy scheduling times.

PS this only affects community instances.

* Better wording

* Mention compute credits limits